backend/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-cors==5.0.1
flask-login==0.6.3
python-dotenv==1.0.1
SQLAlchemy==2.0.39 